+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-05-13
+
 ### Added
 
 -   A method to handle exceptions
@@ -27,5 +29,6 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.1.0...main
+[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.2.0...main
+[0.2.0]: https://github.com/spapanik/pyutilkit/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/spapanik/pyutilkit/releases/tag/v0.1.0